fixing bugs and implementing search bar and catlinks

+++ backend/product-microservice/src/main/java/com/finefoods/productmicroservice/dto/ProductResponse.java
@@ -22,8 +22,8 @@ public class ProductResponse {
     private String tags;
     private String size;
     private String unit;
-    private Integer cost;
-    private Integer currentPrice;
+    private double cost;
+    private double currentPrice;
     private Boolean isTaxed;
     private String skuCode;
     private String upcCode;
